+++ app/src/main/java/com/androidacademy/academyapp2020/view/ui/FragmentMoviesDetails.kt
@@ -1,4 +1,4 @@
-package com.androidacademy.academyapp2020
+package com.androidacademy.academyapp2020.view.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
+import com.androidacademy.academyapp2020.R
 
 class FragmentMoviesDetails : Fragment() {
