--- app/src/main/res/drawable/film_item_gradient.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <gradient
-        android:angle="90"
-        android:endColor="#00191926"
-        android:startColor="#191926" />
-</shape>
