+++ app/src/main/res/drawable/gradient_linear.xml
