+++ app/src/main/res/drawable/gradient_radial.xml
