--- app/src/main/res/drawable/linear_gradient.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <gradient
-        android:angle="90"
-        android:endColor="#80636375"
-        android:startColor="#803B3B4D" />
-    <stroke
-        android:width="1dp"/>
-    <corners
-        android:radius="@dimen/indent_8"/>
-</shape>
