+++ app/src/main/res/layout/activity_main.xml
@@ -4,6 +4,6 @@
     android:id="@+id/fragment_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".view.ui.MainActivity">
 
 </FrameLayout>
