+++ app/src/main/res/layout/fragment_movies_list.xml
@@ -10,7 +10,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/background"
-        tools:context=".FragmentMoviesList">
+        tools:context=".view.ui.FragmentMoviesList">
 
         <View
             android:id="@+id/v_toolbar"
